AcademiaVeritas/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
flask-cors==4.0.0
